[Refactor] MyPage 리팩토링
다른 페이지들과의 통일성을 위해 composition으로 마이페이지의 파일들을 이동,
불필요한 interface를 codegen이 생성해준 interface들로 대체하고 삭제
