--- client/src/composition/FriendList/friend.type.ts
@@ -1,6 +1,0 @@
-export interface IFriend {
-  nickname: string;
-  thumbnail: string;
-  email: string;
-  status: string;
-}
