+++ client/src/composition/MyPage/UserFeedsContainer.tsx
@@ -1,17 +1,13 @@
 import React from 'react';
-import UserFeeds from './UserFeeds';
 import styled from 'styled-components';
+import UserFeeds from './UserFeeds';
 
 const CenterContainer = styled.div`
   margin: 0 auto;
   float: left;
 `;
 
-interface IProps {
-  email: string;
-}
-
-const UserFeedsContainer = ({ email }: IProps) => {
+const UserFeedsContainer = ({ email }: { email: string }) => {
   return (
     <CenterContainer>
       <UserFeeds email={email} />
