--- server/src/api/users/users.graphql
@@ -1,18 +1,0 @@
-type Friend {
-  thumbnail: String
-  nickname: String!
-  email: String!
-  status: String!
-}
-
-type Query {
-  loginUser: Boolean
-  friends: [Friend!]!
-  userWithEmail(email: String!): User
-  friendsWithUserEmail(email: String!): [User]!
-  findRelation(email: String!): String!
-}
-
-type Subscription {
-  updateUserState: Friend!
-}
